TYPE: bug fix
KEYWORDS: string.h warnings

SOURCE: internal

PURPOSE: Remove warnings when building standard.exe because string.h is not included (needed for strncpy).

DESCRIPTION OF CHANGES: 
Add 
#include <string.h> 
to the list of other headers.

LIST OF MODIFIED FILES (annotated if not obvious, not required to be on a single line): 
M	tools/standard.c

TESTS CONDUCTED (explicitly state mandatory, voluntary, and assigned tests, not required to be on a single line):
1) Regression test - do no harm.
2) No more complaints about undefined strncpy.

Index: tools/standard.c
===================================================================
--- tools/standard.c    (revision 9069)
+++ tools/standard.c    (working copy)
@@ -1,5 +1,6 @@
 #include <stdio.h>
 #include <stdlib.h>
+#include <string.h>

 #define LINELEN 8192
 #define STRINGLEN 1024




git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@9088 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d