TYPE: bug fix
KEYWORDS: WPS build io

SOURCE: internal

PURPOSE: Fix the "compile io" option for WRF so that the WPS package may be built without waiting a long time for WRF to finish building.  It used to work, and just fell into disrepair.

DESCRIPTION OF CHANGES: 
The target "toolsdir", which builds the program tools/registry, needed to be added to the rules for building the target "io" (in the top level Makefile).

LIST OF MODIFIED FILES (annotated if not obvious, not required to be on a single line): 
M	Makefile

TESTS CONDUCTED (explicitly state mandatory, voluntary, and assigned tests, not required to be on a single line):
1) Regression test - do no harm
2) The WPS package is now able to build after "compile io" is run.

Index: Makefile
===================================================================
--- Makefile    (revision 9069)
+++ Makefile    (working copy)
@@ -845,7 +845,8 @@
 io :
        @ echo '--------------------------------------'
        ( cd tools ; $(MAKE) standard.exe )
-       ( cd frame ; $(MAKE) io_only )
+       (            $(MAKE) io_only )
+       (            $(MAKE) MODULE_DIRS="$(ALL_MODULES)" toolsdir )
        ( cd frame ; $(MAKE) module_driver_constants.o pack_utils.o module_machine.o module_internal_header_util.o wrf_debug.o )
        ( cd frame ; $(AR) $(ARFLAGS) ../main/libwrflib.a module_driver_constants.o pack_utils.o module_machine.o  \
                                        module_internal_header_util.o module_wrf_error.o wrf_debug.o )



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@9090 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d