TYPE: bug fix
KEYWORDS: tsk_ice, tice_min

SOURCE: internal

PURPOSE: avoid unnecessary model failure

DESCRIPTION OF CHANGES: 
When fractional seaice is used and the fraction is small (but still greater than the minimum of 0.02), there is a chance that the temperature estimated for the seaice portion of the grid becomes too cold (in a case this temp was 24 K), which may cause problems for surface layer physics (in a case, it is MYJSFC)  and potentially other physics too. To avoid this problem, a minimum seaice temp of 221.4 was used if the estimated seaice temp becomes too cold. A related change is to add feedback to variable SSTSK, which should have the same property as SST.

LIST OF MODIFIED FILES:
Registry/Registry.EM_COMMON
phys/module_surface_driver.F

TESTS CONDUCTED:
Reg tested. Tested on the failed user case.



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@9119 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d