TYPE: bug fix
KEYWORDS: real vertical interpolation index check

SOURCE: internal

PURPOSE: Real program fails when running multiple processors with new interpolated chemical data as input (not yet in this version).

DESCRIPTION OF CHANGES:
The horizontal bounding values for the mass point fields (when column-only computations are involved) should always be:
I-Direction: its,MIN(ide-1,ite)
J-Direction: jts,MIN(jde-1,jte)
Going outside this range tripped up the vertical interpolation of the soon-to-be-added chemistry interpolation, as the pressure field was not extended into the halo (as there was no need).  While not harming the other mass point interpolations, it was sloppy indexing that relied on valid input data to exist outside of the specified computational region.

LIST OF MODIFIED FILES (annotated if not obvious, not required to be on a single line):
M       dyn_em/module_initialize_real.F

TESTS CONDUCTED (explicitly state mandatory, voluntary, and assigned tests, not required to be on a single line):
1) Does not run to completion without mods, and successfully completes with mods (using ./configure -D, tested serial and DM on both GNU and Intel)
2) regression test - do no harm
3) real now has a clean bill of health running with bounds checking enabled, serial and DM


Index: module_initialize_real.F
===================================================================
--- module_initialize_real.F	(revision 9122)
+++ module_initialize_real.F	(working copy)
@@ -4389,10 +4389,10 @@
             END IF
          END DO
       ELSE IF ( ( var_type .EQ. 'W' ) .OR.  ( var_type .EQ. 'Z' ) ) THEN
-         istart = MAX(ids  ,its-1)
-         iend   = MIN(ide-1,ite+1)
-         jstart = MAX(jds  ,jts-1)
-         jend   = MIN(jde-1,jte+1)
+         istart = its
+         iend   = MIN(ide-1,ite)
+         jstart = jts
+         jend   = MIN(jde-1,jte)
          kstart = kts
          kend   = kte
          DO j = jstart,jend
@@ -4416,10 +4416,10 @@
             END DO
          END DO
       ELSE IF ( ( var_type .EQ. 'T' ) .OR. ( var_type .EQ. 'Q' ) ) THEN
-         istart = MAX(ids  ,its-1)
-         iend   = MIN(ide-1,ite+1)
-         jstart = MAX(jds  ,jts-1)
-         jend   = MIN(jde-1,jte+1)
+         istart = its
+         iend   = MIN(ide-1,ite)
+         jstart = jts
+         jend   = MIN(jde-1,jte)
          kstart = kts
          kend   = kte-1
          DO j = jstart,jend
@@ -4443,10 +4443,10 @@
             END DO
          END DO
       ELSE
-         istart = MAX(ids  ,its-1)
-         iend   = MIN(ide-1,ite+1)
-         jstart = MAX(jds  ,jts-1)
-         jend   = MIN(jde-1,jte+1)
+         istart = its
+         iend   = MIN(ide-1,ite)
+         jstart = jts
+         jend   = MIN(jde-1,jte)
          kstart = kts
          kend   = kte-1
          DO j = jstart,jend


git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@9123 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d