TYPE: no impact
KEYWORDS: version update

SOURCE: internal

PURPOSE: Update version for pre-release #2 for v3.8.

DESCRIPTION OF CHANGES: 
Modify the character string so that model data now reflects the pre-release #2.

LIST OF MODIFIED FILES (annotated if not obvious, not required to be on a single line): 
M       inc/version_decl

TESTS CONDUCTED (explicitly state mandatory, voluntary, and assigned tests, not required to be on a single line):
1) Cavalierly not running anything at all.


Index: version_decl
===================================================================
--- version_decl	(revision 9178)
+++ version_decl	(working copy)
@@ -1 +1 @@
-   CHARACTER (LEN=10) :: release_version = 'V3.8pre#1 '
+   CHARACTER (LEN=10) :: release_version = 'V3.8pre#2 '


git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@9179 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d