TYPE: bug fix
KEYWORDS: 

SOURCE:Greg Thompson (NCAR)

PURPOSE: Increase the threshold for the number concentration for hail/graupel.  This DIAGNOSTIC is used with the predicted amount of gaupel/hail to give a maximum particle size.  Larger concentration means smaller hail and graupel.

DESCRIPTION OF CHANGES:
Swap the value of a  parameterized constant for a value 10x larger.

LIST OF MODIFIED FILES (annotated if not obvious, not required to be on a single line):
M       phys/module_diag_misc.F

TESTS CONDUCTED (explicitly state mandatory, voluntary, and assigned tests, not required to be on a single line):
1) Greg says this value is better based on a review of real-data cases.

Index: module_diag_misc.F
===================================================================
--- module_diag_misc.F  (revision 9192)
+++ module_diag_misc.F  (working copy)
@@ -251,7 +251,7 @@
 
       DOUBLE PRECISION:: hail_max
       REAL:: hail_max_sp
-      DOUBLE PRECISION, PARAMETER:: thresh_conc = 0.0001d0                 ! number conc. of graupel/hail per cubic meter
+      DOUBLE PRECISION, PARAMETER:: thresh_conc = 0.001d0                  ! number conc. of graupel/hail per cubic meter
       LOGICAL:: scheme_has_graupel
       INTEGER, PARAMETER:: ngbins=50
       DOUBLE PRECISION, DIMENSION(ngbins+1):: xxDx




git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@9200 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d