TYPE: bug fix
KEYWORDS: WRFDA, WPEC, balance_type, registry.var

SOURCE: internal

PURPOSE: Changing the default value of WRFDA namelist variable "balance_type" from 1 to 3

DESCRIPTION OF CHANGES:
According to the contributor of the WPEC capability code (Xin Li), the recommended value of "balance_type" should be 3 (geostrophic AND cyclostrophic terms) as opposed to the current default (1; geostrophic only). Due to this recommendation, changing the default value in registry.var from 1 to 3. This variable is only used in WRFDA (and only used in the WPEC parts of the code) so no changes are expected when this option is not used.

LIST OF MODIFIED FILES:
M       Registry/registry.var

TESTS CONDUCTED: Ran WRFDA regression test for GNU (4.8.2 and 5.3.0) and Intel (12.1.5 and 15.0.3); no unexpected changes or errors. The regression test which enabled the WPEC constraint and did not specify "balance_type" in the namelist showed differences as expected.



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@9216 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d