TYPE: bug fix
KEYWORDS: phys, NoahMP

SOURCE: Michael Barlage (NCAR)

PURPOSE: Mike Barlage sent this explanation for this change:
The cap in noahmp_init is based on an existing hard coded value in the Noah-MP code. The problem I feel we are addressing with this fix is the inconsistency with input sources and the WRF model land cover. The problem I see in both my case where this was a problem and in your case is that SWE and snow depth are being interpolated from adjacent glacier points in ERA-interim. This makes the initial value very large over a non-glacier point in WRF. Some of the points in ERA-I have 10000mm of SWE. This can be something like 35-100m of snow depth (likely not the case for a non-glacier point). I believe the glacier point snow depth/SWE in the NCEP models is lower and thus not an issue.

This hard coded value of 2000mm has come up in other contexts and I asked the users to tell me a more reasonable value, but I have not received one (this may be addressed in the future though). The situation we are trying to avoid is snow building up at a non-glacier point to the level where the surface completely decouples from the atmosphere due to the thermal insulation of the snow pack using a very deep snow layer. The is also due to the three-layer nature of the snow in MP. More layers with deep snow and it wouldn't be as big of a problem.

That said, Noah doesn't care if the snow gets to be 100m and decouples, which is not good, in my opinion.

From the climate simulation perspective, if the snow continues to build up over several years at a non-glacier point, then that point is likely characterized incorrectly in the land cover classification dataset (at least from the WRF perspective). The opposite is true for glacier points that can't sustain snow in the model. Also, though this internal cap is not optimal, the water can still be tracked since any excess is added to the deep runoff.
 
DESCRIPTION OF CHANGES:
Added a check to determine if snow and snowh are consistent and to cap SWE at 2000 mm; increased the length of err_message

LIST OF MODIFIED FILES (annotated if not obvious, not required to be on a single line):
M       phys/module_sf_noahmpdrv.F

TESTS CONDUCTED (explicitly state mandatory, voluntary, and assigned tests, not required to be on a single line): 
1) Ran regression tests for WRFV3 - OK.
2) Code fixes uninitialized / inconsistent snow/snowh

Index: phys/module_sf_noahmpdrv.F
===================================================================
--- phys/module_sf_noahmpdrv.F  (revision 9211)
+++ phys/module_sf_noahmpdrv.F  (working copy)
@@ -1198,7 +1198,7 @@

    INTEGER                   :: errflag, i,j,itf,jtf,ns

-    character(len=80) :: err_message
+    character(len=240) :: err_message
    character(len=4)  :: MMINSL
    character(len=*), intent(in) :: MMINLU
    MMINSL='STAS'
@@ -1226,6 +1226,25 @@
          ENDDO
       ENDIF

+
+       ! Check if snow/snowh are consistent and cap SWE at 2000mm;
+       !  the Noah-MP code does it internally but if we don't do it here, problems ensue
+       DO J = jts,jtf
+          DO I = its,itf
+             IF ( SNOW(i,j) > 0. .AND. SNOWH(i,j) == 0. .OR. SNOWH(i,j) > 0. .AND. SNOW(i,j) == 0.) THEN
+               WRITE(err_message,*)"problem with initial snow fields: snow/snowh>0 while snowh/snow=0 at i,j" &
+                                     ,i,j,snow(i,j),snowh(i,j)
+               CALL wrf_message(err_message)
+             ENDIF
+             IF ( SNOW( i,j ) > 2000. ) THEN
+               SNOWH(I,J) = SNOWH(I,J) * 2000. / SNOW(I,J)      ! SNOW in mm and SNOWH in m
+               SNOW (I,J) = 2000.                               ! cap SNOW at 2000, maintain density
+             ENDIF
+          ENDDO
+       ENDDO
+
       errflag = 0
       DO j = jts,jtf
          DO i = its,itf





git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@9220 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d