TYPE: bug fix
KEYWORDS: <domain> repeat replicate filename

SOURCE: Kevin Manning (NCAR)

PURPOSE: The previous commit that allowed long names and directory structure with repeated information did not allow repeated <domain> string swapping.

DESCRIPTION OF CHANGES: 
The input string to the subroutine is basename.  That is transferred to a
second string (result) where all of the substituions take place.  The first
test (for domain), did not use "result".  Only the first substitution of
<domain> was handled correctly.  The fix is to use "result" as the string that
is manipulated on both the left and right hand sides of the equal sign.

LIST OF MODIFIED FILES (annotated if not obvious, not required to be on a single line): 
M       share/module_io_domain.F

TESTS CONDUCTED (explicitly state mandatory, voluntary, and assigned tests, not required to be on a single line):
1) Regression test - no harm
2) Kevin was successfully able to get a history file name with multiple <domain> string substitutions.

Index: share/module_io_domain.F
===================================================================
--- share/module_io_domain.F	(revision 9288)
+++ share/module_io_domain.F	(working copy)
@@ -204,13 +204,13 @@
   CALL zero_pad ( t1 , fld1 , len1 )
 
 
-! The string name length 12345678  including < >   ----|
-!                        ||||||||                      |
-  i = index( result   , '<domain>' )       !   is this |
-  DO WHILE ( i .GT. 0 )                    !   value   |
-    l = len(trim(result))                  !          \/
-    result = result  (1:i-1) // TRIM(t1) // basename(i+8:l)
-     i = index( result   , '<domain>' ) 
+! The string name length 12345678  including < > ----|
+!                        ||||||||                    |
+  i = index( result   , '<domain>' )       ! is this |
+  DO WHILE ( i .GT. 0 )                    ! value   |
+    l = len(trim(result))                  !        \/
+    result = result  (1:i-1) // TRIM(t1) // result(i+8:l)
+    i = index( result   , '<domain>' ) 
   END DO
 
   i = index( result , '<date>' )



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@9293 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d