TYPE: bug fix
KEYWORDS: moving nest compiler error

SOURCE: internal

PURPOSE: Fix the compile of the ARW moving nest.

DESCRIPTION OF CHANGES: 
An index used by both ARW and NMM for moving nest cases was only declared for
NMM.  When a moving nest is requested, then the "k" integer is declared when
either dynamical core is used.

LIST OF MODIFIED FILES (annotated if not obvious, not required to be on a single line): 
M       share/mediation_nest_move.F

TESTS CONDUCTED (explicitly state mandatory, voluntary, and assigned tests, not required to be on a single line):
1. regression test - no harm
2. built moving nest, ran moving nest case - all OK



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@9337 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d