TYPE: no impact
KEYWORDS: WRFDA, minimization, log files

SOURCE: Zhiquan (Jake) Liu

PURPOSE: Add information to the log printouts from the minimization process, make grepping these values easier.

DESCRIPTION OF CHANGES: In 4DVAR especially, it is tough to visually see the progress of the minimization in the log files due to the number of other messages in the log files. This change adds " minimize_cg" to the beginning of the line that lists the relevant values for each iteration of the minimization (cost function, gradient, step). This way those values can be easily extracted with the grep command to follow the progress of the minimization:

[kavulich@yslogin4 ~/REGTEST]$ grep "^ minimize_cg" radar_4dvar_cv7/rsl.out.0000
 minimize_cg    1    0      6.526203579060955D+03      4.417249774215416D+02      0.000000000000000D+00
 minimize_cg    1    1      4.745589095763956D+03      1.915486767153104D+02      1.825139157513061D-02
 minimize_cg    1    2      3.756573942118002D+03      1.811375336898723D+02      5.391065760500800D-02
 minimize_cg    1    3      3.380797072159185D+03      1.216758478774735D+02      2.290567739693351D-02
 minimize_cg    1    4      3.136399843430361D+03      1.092281768998782D+02      3.301547198253618D-02

This change also adds this information for the first guess (prior to the first minimization step), and adds a column for the "outer loop" value.

LIST OF MODIFIED FILES:
M       var/da/da_minimisation/da_minimise_cg.inc

TESTS CONDUCTED: WRFDA regression tests for gnu and intel show no changes as expected. Log files look as expected as well.



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@9365 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d