TYPE: no impact
KEYWORDS: external program advection kernel

SOURCE: internal

PURPOSE: Upgrade the existing main program that is hidden within the 
advection module source code:
1) allow advect_scalar, advect_scalar_pd, advect_scalar_mono to be used
by the toy/tester main program
2) manufacture a top hat pattern for easy "this is probably OK" testing
3) output info for simple plotting (again, to allow easy identification
of a good advected pattern)
Big picture - the purpose is to allow groups interested in optimizing the
WRF advection to have an easy way to determine if their modifications are
still producing physically meaningful results.

DESCRIPTION OF CHANGES: 
With the same cpp flag (-DADVECT_KERNEL), a few surrounding parts of a main
program are added to the the advection file.  The advect_scalar_pd was already
implemented.  However, to add the other two advection routines, a few extra
lines of "#if ( defined(ADVECT_KERNEL) )" and "#if ( !defined(ADVECT_KERNEL) )"
were required.  This minimizes the impact to the original advection code.

The diffs seem HUGE, but do not be alarmed.  I moved the advection kernel code
(contained within the cpp directives) around so as to not move around any of
the *real* advection routines.  Except for moving around and changing my
own advection kernel bits, just some #if, #elif, and #endif lines have been added.

LIST OF MODIFIED FILES (annotated if not obvious, not required to be on a single line): 
M       module_advect_em.F

TESTS CONDUCTED (explicitly state mandatory, voluntary, and assigned tests, not required to be on a single line):
1) regression test is OK
2) before vs after - no diffs in model output in entire regression test
     Total number of identical file tests: 1800
     Total number of different file tests: 0
3) before vs after - no diffs in the module_advect_em.f90 file
4) advection kernel works as planned



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@9421 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d