TYPE: bug fix
KEYWORDS: WRFDA, CLOUD_CV

SOURCE: internal

PURPOSE: To correct the mz dimension assignment when CLOUD_CV is activated.

DESCRIPTION OF CHANGES:
It is a bug fix, but it does not really have impact.
The bug only affects cloud_cv_options=2, the option that no one
(that I know of) is using, and only under certain circumstances.

mz is an array that stores the dimensions of the vertical
components of control variables. Inside the ifdef CLOUD_CV,
the assigned values to mz(:) are in the order of
v8, v10, v9, and v11. The correct order should be
v8, v9, v10, and v11.
For cloud_cv_options=3, be%v9%mz and be%v10%mz are the same, i.e.
the model vertical levels.

LIST OF MODIFIED FILES:
M       var/da/da_minimisation/da_calculate_j.inc

TESTS CONDUCTED:
WRFDA regtest.
cloud_cv_options=2 is not in the regtest.
The change has no impact on other current tests.



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@9516 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d