SOURCE: internal
PURPOSE: Clean up the wrong "type" of initialization in Registry and various READMEs.  If a variable is an integer, 
for example, then 0.0 is not a valid value for an initialization in the Registry.  

DESCRIPTION OF CHANGES: 
In the Registry directory, we can find all of the integer variables in the 
namelist that are initialized as a floating point values.
grep -i rconfig [Rr]eg* | grep -i integer | awk '{print $3, $6}' | grep "\."
dust_dsr 0.
dust_veg 0.
dust_soils 0.
dust_smois 0.
scm_canwat 0.0
fire_num_ignitions 0.
rtminit_satid -1.0
rtminit_sensor -1.0
All of the above have been de-decimalized.

I looked to see if there were real namelist entries that were initialized as integers.  This is not really a
problem, but for consistency it should probably be fixed.  The problem is that there are a LOT of these:
grep -i rconfig [Rr]eg* | grep -i real | grep -vi integer | grep -vi character | awk '{print $3, $6}' | grep -v "\." | wc -l
140
Therefore, no effort is being made to address this "real, but initialized as an integer within the Registry" issue.

In the run and test/em_real directories, the values for ocean_z (layer depths of the ocean)
in the README and example files are now shown with a decimal point.

Lastly, one of the variables that was initialized with 0.0 (and declared an integer) should be a real.
find . -name \*.F -exec grep -i scm_canwat {} \; -print
              grid%canwat(i,j) = config_flags%scm_canwat
./dyn_em/module_initialize_scm_xy.F

grep -i " canwat " Registry/Registry.EM_COMMON
state    real   CANWAT           ij     misc        ...

LIST OF MODIFIED FILES (annotated if not obvious, not required to be on a single line): 
M       Registry/Registry.EM_COMMON
M       Registry/registry.chem
M       Registry/registry.fire
M       Registry/registry.var
M       run/README.namelist
M       test/em_real/examples.namelist

TESTS CONDUCTED (explicitly state mandatory, voluntary, and assigned tests, not required to be on a single line):
1) Regression tests, passes as expected



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@9520 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d