TYPE: bug fix
KEYWORDS: max wind tropopause GFS real interpolation

SOURCE: internal 

PURPOSE: remove spurious values introduced with vertical interpolation, reported 
by Jim Bresch (NCAR)

DESCRIPTION OF CHANGES: 
This is a problem caused by the use of a high-order interpolator (in this case a cubic
spline) with the introduction of additional data surfaces that are near an existing
isobaric surface (the GFS extra fields for the max wind level, and the trop level).

There are two ways to deal with this:

FIRST METHOD THAT IS INCLUDED, BUT IFDEF'ED OUT
===============================================
The solution is to run through each column before interpolation to verify that no
pressures are closer than the user-specified zap_close_levels.  If two neighboring 
levels are found that are "close", a level is removed and all of the data "above" in 
the 1d array is brought down an index.

There are a couple of safety nets:
1) The zap_close_levels is user specified.  The user can make this quite small if
no additional levels are added, and the user has really high vertical resolution.
2) The levels are not removed where the pressure levels are small.  For example, the
top of the GFS has pressure levels of 1, 2, 3, 5, 7, 10 mb.  A 5 mb zap_close_levels
(the default) would not work.  The actual test requires that the pressure level being
removed has to be 10x the size of the zap_close_levels.
3) If more than two levels are removed, the code stops.  The user will not accidentally
get only data at 1000 mb and the model lid if all of the levels are 1 mb apart.

With more testing, this could become an active and executable part of the code.


SECOND METHOD THAT IS NOW INSERTED IN THE CODE
================================================
Try to detect if a user has pressure levels too close together, and then report
that informtion to the user.  In Jim's test case (with two really bad places in the
V variable), this simple test is able to find both of them, and only detected these
two points:

*** -> Check your wrfinput_d01 file, you might have input pressure levels too close together ( 15007.46 Pa and  14999.93 Pa) at ( 201, 255) for variable type V
*** -> Check your wrfinput_d01 file, you might have input pressure levels too close together ( 20005.96 Pa and  19999.92 Pa) at ( 146, 275) for variable type V

When a user actually has any troubles that can be attributed to bad vertical interpolation 
with additional levels (as indicated above), the wrfhelpers can suggest several things:
1) make the interpolation linear
&domains
 lagrange_order                      = 1,
/
2) do not use the max wind or tropopause-level data
&domains
 use_maxw_level                      = 0,
 use_trop_level                      = 0,
/
3) un-ifdef the mod in the vertical interpolation routine (look for #if 0 in
the vertical interpolation routine)


LIST OF MODIFIED FILES (annotated if not obvious, not required to be on a single line): 
M      dyn_em/module_initialize_real.F

TESTS CONDUCTED (explicitly state mandatory, voluntary, and assigned tests, not required to be on a single line):
1) Regression test OK
2) Jim's case with 1200 m/s V, now is a reasonable 19 m/s
3) The results with the cubic spline are similar enough to the linear interpolation
4) POD and FAR both acceptable for this simple delta-pressure check



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@9522 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d