Fix an error introduced in the 42449c23 commit.
The setting of calc_wpec_term should be outside of the if ( .not. var4d ) block,
otherwise it is uninitialized when var4d=true and then it is used later
for determinging if certain calculation and halo exchange should be done or not.
The failure behavior will be compiler and compiler version dependent.