educe computational patch size for intermediate domain for vertical nesting
TYPE: bug fix

KEYWORDS: vertical refinement, intermediate domain, patch

SOURCE: internal

DESCRIPTION OF CHANGES: One or more paragraphs describing problem, solution, and required changes.
The vertical refinement option rebalances the fine grid to
account for the horizontal and vertical interpolation.  The
intermediate domain value are coupled with total column pressure.
When the allocated but not initialized portions of the patch
are part of the computations, floating point errors show up
when using the FPE traps.

The solution is to restrict the patch dimensions to be entirely
enclosed with the domain dimensions (for the intermediate
domain).

LIST OF MODIFIED FILES: list of changed files (use `git diff --name-status master` to get formatted list)
external/RSL_LITE/module_dm.F

TESTS CONDUCTED: Explicitly state if a WTF and or other tests were run, or are pending. For more complicated changes please be explicit!
1) With the mods, the code successfully runs through several time
steps with the FPE flags activated.
2) TO DO: standard WTF
3) TO DO: compare results from modified version to pristine master,
they should be the same