Intermediate domain DO index bounds for vert refinement rebalance
Mods to fix the incorrect intermediate domain indexing for the
vertical refinement code (during the rebalancing that occurs
for each parent time step).  The horizontal indexing for the
intermediate domain should not be ips->ipe, but the more
restrictive MAX(ids,ips)->MIN(ide-1,ipe) ... and of course for
the J direction, too.  This will be committed to the master
soon, but we need this in the hybrid coordinate merge branch
to allow continued testing.

modified:   external/RSL_LITE/module_dm.F