Reduce computational patch size for intermediate domain for vertical nesting (#37)
TYPE: bug fix

KEYWORDS: vertical refinement, intermediate domain, patch

SOURCE: internal

DESCRIPTION OF CHANGES:
The vertical refinement option rebalances the fine grid to
account for the horizontal and vertical interpolation. The
intermediate domain value are coupled with total column pressure.
When the allocated (but initialized to zero) portions of the patch
are part of the computations, floating point errors show up
when using the FPE traps. These values are never used, but the
uninitialized values throw FPEs when we are doing division.

The solution is to require that the computations be entirely
enclosed with the more restrictive of either the patch or domain dimensions
(for the intermediate domain, "patch" and "domain" sizes are not
as intuitive as on a parent or child domain).

LIST OF MODIFIED FILES:
external/RSL_LITE/module_dm.F

TESTS CONDUCTED:

1) With the mods, the vertical refinement code successfully runs through several time
steps with the FPE flags activated, and the code was not able to
do so without these mods.
2) WTF passes (including the several dozen vertical refinement tests)
3) The results (with the restrictive index bounds) are bit-for-bit identical with the master results (the restricted index computations are INDEED giving the same results). This
is a comparison of the 1798 regression tests (with and without vertical
nesting).