Refix tendency calculation and correct height calculation in NSAS (#105)
TYPE: bug fix (revised)

KEYWORDS: NSAS cumulus scheme, tendencies, shallow convection

SOURCE: Kun Liu of CMA, China and Jihyeon Jang, NCAR

DESCRIPTION OF CHANGES:
This commit contains two pieces: the first piece is a revision of the bug fix proposed by Sebastien Masson. Jihyeon's fix is to pass delt (which is the specified cumulus step) to both deep and shallow schemes, instead of dt (which is the dynamic time step). This will make the tendency calculations consistent. And the previous fix failed to fix the rainfall tendency. This change doesn't change the tendencies much, but recovers lost rainfall.

The second piece is to fix some indexing when height field is used. This has a small impact in the test conducted.

LIST OF MODIFIED FILES:
phys/module_cu_nsas.F

TESTS CONDUCTED: reg tested. A pair of tests are conducted to compare before and after changes in a 45 km domain. The changes are small in these tests.