Fix failures for IASI radiance assimilation with PGI compiler (#109)
TYPE: bug fix

KEYWORDS: WRFDA, IASI, PGI

SOURCE: internal

DESCRIPTION OF CHANGES: In the da_qc_iasi routine, there is an extra, ambiguous "close(iunit)" statement at the end. Not only is this not paired with an equivalent "open(iunit)" statement, but iunit is not even defined in this subroutine! Yet, for some strange reason, the only compiler I can find that fails due to this is PGI 15.1.

Removing the rogue "close" statement fixes the error. In addition, I took this opportunity to remove a number of other variables in this subroutine that are declared but never defined or used, presumably due to copy-pasting of another routine as a template.

LIST OF MODIFIED FILES:
M       var/da/da_radiance/da_qc_iasi.inc

TESTS CONDUCTED: WRFDA regtest compiled with PGI 15.1 failed before this fix for IASI assimilation test. This fix, on top of the fix in PR #108 (8ddf96a), allows the test to run successfully. gfortran and intel tests did not exhibit this error, and show no impact. WRFDA Regression tests now all pass for PGI 15.1! (except for one known 4DVAR failure)