Fix non-standard format strings in setup_physics_suite()
The setup_physics_suite() routine previously used format strings like
'(A21,*(I6,A1))' with the intention of repeating the (I6,A1) part some
number of times determined at runtime. The use of '*' as a repetition
count isn't valid in Fortran 90/95/2003, however, and was caught by
the PGI compiler.

This commit introduces code to build a format statement based on the
runtime-determined repetition count, and to then use that format statement
when printing out the schemes selected in a physics suite.