Set default physics suite to 'none'
This commit sets the default physics suite to 'none', and adds code to print
a simple message and return early from the suite setup routine if the suite
is 'none'. This should avoid any confusing output in the rsl files for anyone
using existing namelist.input files with WRF v3.9.