Enable initialization of ARW with QV (mixing ratio) (#163)
TYPE: enhancement

KEYWORDS: initialization mixing ratio qv

SOURCE: internal

PURPOSE:
Enable initialization of the QVAPOR field in ARW from a first-guess
QV (mixing ratio) field

DESCRIPTION OF CHANGES:
This PR adds a block of code to the init_domain_rk( ) routine to handle the case
when the FLAG_QV variable is set to 1, thereby allowing the initialization of
the RH field from the first-guess QV (mixing ratio) field. The RH field is then
interpolated and subsequently used to initialize the model QVAPOR field as
usual. The changes for using QV are largely copy-and-paste from the code to
use SPECHUMD.

LIST OF MODIFIED FILES:
* dyn_em/module_initialize_real.F

TESTS CONDUCTED:
1. All WTF regression tests pass except for several unexplained BFB failures
   with PGI.
2. Initialization of a WRF simulation using MPAS output data, which provides
   only mixing ratio as the moisture field, produces plausible initial and
   boundary values for the QVAPOR field.