Collect all fatal errors in input_wrf for better user interface
TYPE: enhancement

KEYWORDS: fatal, error, input_wrf

SOURCE: internal

DESCRIPTION OF CHANGES:
The good example is the file module_check_a_mundo.F. In that file, the namelist settings are tested. When fatal problems are found, a flag is set that at least one fata error has occured. At the end of the file, after all of the error messages have been printed out, THEN the fatal error is issued. Now we do the same inside of the input_wrf.F file.

There are flags to count the errors and to also have a boolean setting whether a fatal error should be issued. At the end of the file the fatal error is handled with the usual wrf_error_fatal call.

LIST OF MODIFIED FILES:
M       share/input_wrf.F

TESTS CONDUCTED:
- [ ] A few spot tests to see if things are correctly detected.
- [ ] Reggie 3.05