Real and WRF require consistent gwd_opt settings
TYPE: bug fix

KEYWORDS: gwd gwd_opt

SOURCE: internal, problem found by Jim Bresch (NCAR)

DESCRIPTION OF CHANGES:
There are only a handful of settings related to physical parameterization settings in the WRF model that need to be set consistently when running the real program. The sf_surface_physics and sst_update options are examples. Another is gwd_opt. When gwd_opt is not set, it has the default value of zero. With that value, the two fields (OA and OL) are not in the wrfinput_dxx file that the WRF model reads. If gwd_opt is then turned on in the WRF model, those absent (and therefore zeroed out) fields allow the model to complete "successfully", but with no changes in results.

LIST OF MODIFIED FILES:
M       share/input_wrf.F
M       share/output_wrf.F

TESTS CONDUCTED:
- [ ] Test new settings, problematic nml detected with gwd_opt
- [ ] Reggie, 3.05

REDUNDANT stuff from PR #177 (so I can test) is temporarily included