New script to help determine what is wrong with a namelist.input file
TYPE: new feature

KEYWORDS: namelist, column, max_dom

SOURCE: internal, suggested by Sebastien Masson

DESCRIPTION OF CHANGES:
New script that checks the user's namelist.input file for consistency with the Registry.
1. Is the namelist variable in one of the Registry files
2. For single entry values, is there more than one column in the namelist.input file
3. For max_domains entry values, are there at least max_dom entries listed

There are some gotchas, which is why this is recommended as an in-house tool for now. There are some namelist entries that are automatically manufactured (such as auxinput4 items), and there are some namelist values that are not defined in any registry (quilting). Before release to general users, these scenarios would need to be handled. However, for the wrfhelp folks, the information is easily understood.

LIST OF MODIFIED FILES: list of changed files (use `git diff --name-status master` to get formatted list)
M       tools/commit_nml_check.txt

TESTS CONDUCTED: Explicitly state if a WTF and or other tests were run, or are pending. For more complicated changes please be explicit!
1. Picks up obvious problems in test namelists
2. Mostly for use by wrfhelp, not individual users